Add files via upload
Write a function cycle_convert that takes a variable x as an input and a cycle step size n. It
converts the type of x into another type and returns the converted x. The conversion is in the
following fixed order of Python types:
int → float → bool → string → complex
Further instructions:
• Cycle step size n controls how many steps you go. For instance, for n=1, an int is
converted to float, a string is converted to complex, and so on. For n=2, an int is
converted to bool (skipping the conversion to float), or a float is converted to a string
(skipping the conversion to bool).
• Sometimes conversion is not possible. For instance, you cannot convert random words
to numbers, but you needn’t worry about this. Assume all the conversions are
possible.
• Wrap around: When the step size brings you beyond the end of the conversion order,
you start from the beginning. For instance, for n=2, a complex type is converted to
float. Or for n=4, a string is converted to bool.
• Negative step size: if n negative, the cycle goes to the left. For instance, for n=-1, a
complex type is converted to a string, or a bool type is converted to float.
• The default value for n is 1.