Pandas & Visualisation
You are given the CSV file orders_cardiff_drinks.csv with a few thousand sales made by a company called Cardiff Drinks. This company sells wine, soda, and juice. The CSV file has information about the order id, the client id, when the order was made, the quantity (qty) of each of the three products the company sells (soda, wine, and juice), and the money spent on each product (in GBP), and the totals for quantity and money. Read this file using Pandas' read_csv() method and answer the following questions:

Q2.1 (visualization). Produce a bar plot that shows, for each year, the quantity sold for each product. 
Q2.2 (visualization). Produce a correlation matrix and heatmap visualization between GBP and qty values for soda, juice, and wine.
Q2.3 (answering questions). To answer the following questions, write Python code for querying the data frame you have loaded with read_csv():
• How many orders were placed in 2016?
• Which client has purchased the greatest number of soda bottles overall?
• On which weekday is wine sold the most on average (across the whole sales history)?
• Which drink was the most popular (in GBP) in winter (from Dec 21st to March 21st)?
• In which year-month (e.g. January 2015) was there the highest number of orders in which the number of wine bottles was higher than the average for that year?