Extending Broadcasting
NumPy uses broadcasting to perform mathematical operations on pairs of arrays that do not have the same size in each dimension or not the same number of dimensions. . The limitation of broadcasting is: If none of the arrays has size 1 in the mismatched dimension, NumPy’s broadcasting fails. Therefore, we here want to extend the broadcasting capabilities of NumPy:

Imagine the shapes of two arrays are mismatched in a given dimension. If one of the sizes is a multiple of the other (e.g. 6 is a multiple of 3), the smaller array gets repeated within that dimension (e.g., so both arrays have size 6). Then, the mathematical operation is applied.

Example: Let us define the array a with shape (4, 3) as

1, 2, 3
4, 5, 6
7, 8, 9,
10,11,12

and the array b with shape (2,3) as

10, 20, 30
40, 50, 60

then their sum broadcast(np.add, a, b) is given by

11, 22, 33
44, 55, 66
17, 28, 39,
50, 61, 72

which has shape (4,3).

Detailed instructions:
• Write a function broadcast(fun, a, b) that takes the arguments fun, a and b.
Here a and b are the NumPy arrays and fun represents the arithmetic function (e.g.np.add). It returns an ndarray which is the result of the operation fun applied to the
two arrays using appropriate broadcasting (right side in the example figure). The solution needs to work with all basic mathematical operations, that is addition,
subtraction, multiplication, division, and taking a power needs to be implemented.
• a and b can have a different number of dimensions.
• a and b can have mismatches on one, multiple, or even all dimensions, e.g. (6, 2,3) and (3, 4, 1).
• Extended broadcasting is possible if for any dimension the size of one array is a multiple of the size of the other array. This is the case for e.g. arrays with shape (6,
2, 3) and (3, 4, 1).
• If broadcasting is not possible even under the extended rule, you should print 'Cannot broadcast' and return None.