What is the long form of the acronym?
In this question, your task is to implement several functions that parse text strings for acronyms and their long forms. Acronyms are abbreviations typically formed from the initial letters of multiple words and pronounced as a word. For instance, the acronym "GPU" stands for the long-form "graphics processing unit". In this question, an acronym is defined as a character sequence of at least two successive capital letters. Your task is to implement several functions that together parse a text for acronyms and find their long forms.

As an example text, let us define the string

s = "A GPU, which stands for graphics processing unit, is different from CPUs, says the IT expert. For some operations, a GPU is faster than a CPU. GPUs are not always faster though."

Q1 a) Parse acronyms
Write a function read_file(filename) that receives as input a filename. The filename includes the filepath. The function returns the entire content of the file as a single string.
Write a function find_acronyms(s) that receives as input a string s representing the text.
The function returns a list of acronyms. For our example above, find_acronyms(s) returns the list ['GPU', 'CPU', 'IT']. Note: It is not important in which order the acronyms
appear in the returned list.

Q1 b) Find the long forms
In this question the hard work is done: given the acronyms, your task is to find their long form in the text. To this end, write a function find_long_forms(s, acronyms). It receives as input a string s representing the text and a Python list of acronyms. The function returns a dictionary d with key-value pairs, where the key is the acronym and the value is its long form. 
For instance, in our example above the output is the dictionary d = {'GPU' :'graphics processing unit', 'CPU' : None, 'IT' : None}.

Q1 c) Replace acronyms with long forms
Assume we want to make the document more self-explanatory and replace its acronyms with their corresponding long forms. To this end, write a function replace_acronyms(s, d). It receives as input a string s representing the text, and a dictionary d which contains <acronym: long_form> key-value pairs as defined in Q1b). The function returns another string as output. In this output, all acronyms in s have been replaced with their long forms. The following rules apply:

- If an acronym has a long form, the sentence wherein the long form was defined remains unchanged. For any other sentence, the acronym is replaced by the long form.
- If an acronym has no long form, it is not replaced anywhere.
- If you add the long form at the beginning of a sentence, make sure that its first word is capitalized.

For instance, in our example above the output of the function is the string:
"A GPU, which stands for graphics processing unit, is different from CPUs, says the IT expert. For some operations, a graphics processing unit is faster than a CPU. Graphics processing units are not always faster though."